types定義の整理
