+++ client/src/contexts/group-context.tsx
@@ -1,5 +1,5 @@
+import { Group } from "@/types/common";
 import { createContext } from "react";
-import { Group } from "@/types/groupData";
 
 interface GroupContextType {
   groups: Group[];
@@ -10,5 +10,5 @@ interface GroupContextType {
 }
 
 export const GroupContext = createContext<GroupContextType | undefined>(
-  undefined,
+  undefined
 );
