--- client/src/types/groupData.ts
@@ -1,10 +1,0 @@
-export type Group = {
-  groupName: string;
-  memberCount: number;
-  groupImage: string;
-  groupId: string;
-};
-
-export type GroupData = {
-  groupData: Group[];
-};
