+++ client/src/types/layout.ts
@@ -1,0 +1,25 @@
+export interface LayoutProps {
+  version: string;
+  avatar: string;
+  children: React.ReactNode;
+  onLogout: () => void;
+}
+
+export interface HeaderProps {
+  avatar: string;
+  onSidebar: () => void;
+}
+
+export interface HomeAvatarProps {
+  src: string;
+}
+
+export interface SidebarProps {
+  version: string;
+  onLogout: () => void;
+}
+
+export interface SidebarListItemProps {
+  text: string;
+  onClick: () => void;
+}
