--- client/src/types/loading.ts
@@ -1,3 +1,0 @@
-export interface LoadingIndicatorProps {
-  message?: string;
-}
